+++ HISTORY.md
@@ -10,6 +10,8 @@
 
 #### Improvements
 
+- Use chunk computation in CVT brute force calculation to reduce memory usage
+  ({pr}`394`)
 - Test pyribs installation in tutorials ({pr}`384`)
 - Add cron job for testing installation ({pr}`389`)
 - Fix broken cross-refs in docs ({pr}`393`)